requirements.txt
numpy
pandas
python-dotenv
langchain
langchain-groq
langchain_community
faiss-cpu
pypdf
langchain_huggingface
sentence-transformers
ipykernel
wikipedia
numexpr 
Django 